+++ app/event_source/events/iap/applications/magi_medicaid_application_determined.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Events
+  module Iap
+    module Applications
+      # This class will register event
+      class MagiMedicaidApplicationDetermined < EventSource::Event
+        publisher_path 'publishers.application_determined_publisher'
+
+      end
+    end
+  end
+end
