+++ app/event_source/publishers/application_determined_publisher.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module Publishers
+  # Publisher will send request payload to medicaid gateway for determinations
+  class ApplicationDeterminedPublisher
+    include ::EventSource::Publisher[amqp: 'enroll.fdsh.verifications']
+
+    register_event 'magi_medicaid_application_determined'
+  end
+end
